work on probe
