+++ refl1d/probe.py
@@ -12,8 +12,6 @@ See :ref:`data-guide` for details.
 
 """
 
-from __future__ import with_statement, division, print_function
-
 # TOF stitching introduces a lot of complexity.
 # Theta offset:
 #   Q1 = 4 pi sin(T1 + offset)/L1
@@ -55,7 +53,6 @@ from bumps.util import field, field_desc, Optional, Any, Union, Dict, Callable, 
 
 from periodictable import nsf, xsf
 from bumps.parameter import Parameter, to_dict, Constant
-from bumps.plotutil import coordinated_colors, auto_shift
 from bumps.data import parse_multi, strip_quotes
 from bumps.util import USE_PYDANTIC
 
@@ -346,261 +343,6 @@ class BaseProbe:
             # print("saving", A)
             fid.write(asbytes(header))
             np.savetxt(fid, A.T, fmt="%20.15g")
-
-    def plot(self, view=None, **kwargs):
-        """
-        Plot theory against data.
-
-        Need substrate/surface for Fresnel-normalized reflectivity
-        """
-
-        view = view if view is not None else self.view
-
-        if view == "linear":
-            self.plot_linear(**kwargs)
-        elif view == "log":
-            self.plot_log(**kwargs)
-        elif view == "fresnel":
-            self.plot_fresnel(**kwargs)
-        elif view == "logfresnel":
-            self.plot_logfresnel(**kwargs)
-        elif view == "q4":
-            self.plot_Q4(**kwargs)
-        elif view == "resolution":
-            self.plot_resolution(**kwargs)
-        elif view.startswith("resid"):
-            self.plot_residuals(**kwargs)
-        elif view == "fft":
-            self.plot_fft(**kwargs)
-        elif view == "SA":  # SA does not plot since it does not exist
-            pass
-        else:
-            raise TypeError("incorrect reflectivity view '%s'" % view)
-
-    def plot_resolution(self, suffix="", label=None, **kwargs):
-        import matplotlib.pyplot as plt
-
-        plt.plot(self.Q, self.dQ, label=self.label(prefix=label, suffix=suffix))
-        plt.xlabel(r"Q ($\AA^{-1}$)")
-        plt.ylabel(r"Q resolution ($1-\sigma \AA^{-1}$)")
-        plt.title("Measurement resolution")
-
-    def plot_linear(self, **kwargs):
-        """
-        Plot the data associated with probe.
-        """
-        import matplotlib.pyplot as plt
-
-        self._plot_pair(ylabel="Reflectivity", **kwargs)
-        plt.yscale("linear")
-
-    def plot_log(self, **kwargs):
-        """
-        Plot the data associated with probe.
-        """
-        import matplotlib.pyplot as plt
-
-        self._plot_pair(ylabel="Reflectivity", **kwargs)
-        plt.yscale("log")
-
-    def plot_logfresnel(self, *args, **kw):
-        """
-        Plot the log Fresnel-normalized reflectivity associated with the probe.
-        """
-        import matplotlib.pyplot as plt
-
-        self.plot_fresnel(*args, **kw)
-        plt.yscale("log")
-
-    def plot_fresnel(self, substrate=None, surface=None, **kwargs):
-        r"""
-        Plot the Fresnel-normalized reflectivity associated with the probe.
-
-        Note that the Fresnel reflectivity has the intensity and background
-        applied before normalizing so that hydrogenated samples display more
-        cleanly.  The formula to reproduce the graph is:
-
-        .. math::
-
-             R' = R / (F(Q) I + B)
-             \Delta R' = \Delta R / (F(Q) I + B)
-
-        where $I$ is the intensity and $B$ is the background.
-        """
-        if substrate is None and surface is None:
-            raise TypeError("Fresnel-normalized reflectivity needs substrate or surface")
-        F = self.fresnel(substrate=substrate, surface=surface)
-
-        # print("substrate", substrate, "surface", surface)
-        def scale(Q, dQ, R, dR, interpolation=0):
-            Q, fresnel = self.apply_beam(self.calc_Q, F(self.calc_Q), interpolation=interpolation)
-            return Q, dQ, R / fresnel, dR / fresnel
-
-        if substrate is None:
-            name = "air:%s" % surface.name
-        elif surface is None or isinstance(surface, Vacuum):
-            name = substrate.name
-        else:
-            name = "%s:%s" % (substrate.name, surface.name)
-        self._plot_pair(scale=scale, ylabel="R/(R(%s)" % name, **kwargs)
-
-    def plot_Q4(self, **kwargs):
-        r"""
-        Plot the Q**4 reflectivity associated with the probe.
-
-        Note that Q**4 reflectivity has the intensity and background applied
-        so that hydrogenated samples display more cleanly.  The formula
-        to reproduce the graph is:
-
-        .. math::
-
-             R' = R / ( (100*Q)^{-4} I + B)
-             \Delta R' = \Delta R / ( (100*Q)^{-4} I + B )
-
-        where $B$ is the background.
-        """
-
-        def scale(Q, dQ, R, dR, interpolation=0):
-            # Q4 = np.maximum(1e-8*Q**-4, self.background.value)
-            Q4 = 1e-8 * Q**-4 * self.intensity.value + self.background.value
-            return Q, dQ, R / Q4, dR / Q4
-
-        # Q4[Q4==0] = 1
-        self._plot_pair(scale=scale, ylabel="R (100 Q)^4", **kwargs)
-
-    def _plot_pair(
-        self,
-        theory=None,
-        scale=lambda Q, dQ, R, dR, interpolation=0: (Q, dQ, R, dR),
-        ylabel="",
-        suffix="",
-        label=None,
-        plot_shift=None,
-        **kwargs,
-    ):
-        import matplotlib.pyplot as plt
-
-        c = coordinated_colors()
-        plot_shift = plot_shift if plot_shift is not None else Probe.plot_shift
-        trans = auto_shift(plot_shift)
-        if hasattr(self, "R") and self.R is not None:
-            Q, dQ, R, dR = scale(self.Q, self.dQ, self.R, self.dR)
-            if not self.show_resolution:
-                dQ = None
-            plt.errorbar(
-                Q,
-                R,
-                yerr=dR,
-                xerr=dQ,
-                capsize=0,
-                fmt=".",
-                color=c["light"],
-                transform=trans,
-                label=self.label(prefix=label, gloss="data", suffix=suffix),
-            )
-        if theory is not None:
-            # TODO: completely restructure interpolation handling
-            # Interpolation is used to show the theory curve between the
-            # data points.  The _calc_Q points used to predict theory at
-            # the measured data are used for the interpolated Q points, with
-            # the resolution function centered on each interpolated value.
-            # The result is that when interpolation != 0, there are more
-            # theory points than data points, and we will need to accomodate
-            # this when computing normalization curves for Fresnel and Q^4
-            # reflectivity.
-            # Issues with current implementation:
-            # * If the resolution is too tight, there won't be sufficient
-            #   support from _calc_Q to compute theory at Q interpolated.
-            # * dQ for the interpolated points uses linear interpolation
-            #   of dQ between neighbours.  If measurements with tight and
-            #   loose resolution are interleaved the result will look very
-            #   strange.
-            # * There are too many assumptions about interpolation shared
-            #   between Experiment and Probe objects.  In particular, the
-            #   Fresnel object needs to be computed at the same interpolated
-            #   points as the theory function.
-            # * If there are large gaps in the data the interpolation will
-            #   not fill them in correctly.  Perhaps we should set _Q_plot
-            #   and _calc_Q_plot independently from the data?
-            # * We sometimes show the theory curve without resolution
-            #   applied; this has not been tested with interpolation
-            interpolation = kwargs.get("interpolation", 0)
-            Q, R = theory
-            Q, dQ, R, dR = scale(Q, 0, R, 0, interpolation=interpolation)
-            plt.plot(
-                Q,
-                R,
-                "-",
-                color=c["dark"],
-                transform=trans,
-                label=self.label(prefix=label, gloss="theory", suffix=suffix),
-            )
-            # from numpy.fft import fft
-            # x, y = Q[1::2], abs(fft(R)[:(len(R)-1)//2])
-            # y = y * (R.max()/y.max())
-            # plt.plot(x, y, '-')
-        plt.xlabel("Q (inv Angstroms)")
-        plt.ylabel(ylabel)
-        # plt.legend(bbox_to_anchor=(1.05, 1), loc=2, borderaxespad=0.)
-        h = plt.legend(fancybox=True, numpoints=1)
-        h.get_frame().set_alpha(0.5)
-
-    def plot_residuals(self, theory=None, suffix="", label=None, plot_shift=None, **kwargs):
-        import matplotlib.pyplot as plt
-
-        plot_shift = plot_shift if plot_shift is not None else Probe.residuals_shift
-        trans = auto_shift(plot_shift)
-        if theory is not None and self.R is not None:
-            c = coordinated_colors()
-            Q, R = theory
-            # In case theory curve is evaluated at more/different points...
-            R = np.interp(self.Q, Q, R)
-            residual = (R - self.R) / self.dR
-            plt.plot(
-                self.Q, residual, ".", color=c["light"], transform=trans, label=self.label(prefix=label, suffix=suffix)
-            )
-        plt.axhline(1, color="black", ls="--", lw=1)
-        plt.axhline(0, color="black", lw=1)
-        plt.axhline(-1, color="black", ls="--", lw=1)
-        plt.xlabel("Q (inv A)")
-        plt.ylabel("(theory-data)/error")
-        plt.legend(numpoints=1)
-
-    def plot_fft(self, theory=None, suffix="", label=None, substrate=None, surface=None, **kwargs):
-        """
-        FFT analysis of reflectivity signal.
-        """
-        raise NotImplementedError
-        import matplotlib.pyplot as plt
-
-        c = coordinated_colors()
-        if substrate is None and surface is None:
-            raise TypeError("FFT reflectivity needs substrate or surface")
-        F = self.fresnel(substrate=substrate, surface=surface)
-        # Qc = sqrt(16*pi*substrate)
-        Qc = 0
-        Qmax = max(self.Q)
-        T = np.linspace(Qc, Qmax, len(self.Q))
-        z = np.linspace(0, 2 * pi / Qmax, len(self.Q) // 2)
-        if hasattr(self, "R"):
-            signal = np.interp(T, self.Q, self.R / F(self.Q))
-            A = abs(numpy.fft.fft(signal - np.average(signal)))
-            A = A[: len(A) // 2]
-            plt.plot(z, A, ".-", color=c["light"], label=self.label(prefix=label, gloss="data", suffix=suffix))
-        if theory is not None:
-            Q, R = theory
-            signal = np.interp(T, Q, R / F(Q))
-            A = abs(numpy.fft.fft(signal - np.average(signal)))
-            A = A[: len(A) // 2]
-            plt.plot(z, A, "-", color=c["dark"], label=self.label(prefix=label, gloss="theory", suffix=suffix))
-        plt.xlabel("w (A)")
-        if substrate is None:
-            name = "air:%s" % surface.name
-        elif surface is None or isinstance(surface, Vacuum):
-            name = substrate.name
-        else:
-            name = "%s:%s" % (substrate.name, surface.name)
-        plt.ylabel("|FFT(R/R(%s))|" % name)
 
     def label(self, prefix=None, gloss="", suffix=""):
         if not prefix:
@@ -1207,56 +949,6 @@ class ProbeSet(Probe):
             p.save(filename + str(i + 1), th, substrate=substrate, surface=surface)
 
     save.__doc__ = Probe.save.__doc__
-
-    def plot(self, theory=None, **kw):
-        import matplotlib.pyplot as plt
-
-        for p, th in self.parts(theory):
-            p.plot(theory=th, **kw)
-
-    plot.__doc__ = Probe.plot.__doc__
-
-    def plot_resolution(self, **kw):
-        for p in self.probes:
-            p.plot_resolution(**kw)
-
-    plot_resolution.__doc__ = Probe.plot_resolution.__doc__
-
-    def plot_linear(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_linear(theory=th, **kw)
-
-    plot_linear.__doc__ = Probe.plot_linear.__doc__
-
-    def plot_log(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_log(theory=th, **kw)
-
-    plot_log.__doc__ = Probe.plot_log.__doc__
-
-    def plot_fresnel(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_fresnel(theory=th, **kw)
-
-    plot_fresnel.__doc__ = Probe.plot_fresnel.__doc__
-
-    def plot_logfresnel(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_logfresnel(theory=th, **kw)
-
-    plot_logfresnel.__doc__ = Probe.plot_logfresnel.__doc__
-
-    def plot_Q4(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_Q4(theory=th, **kw)
-
-    plot_Q4.__doc__ = Probe.plot_Q4.__doc__
-
-    def plot_residuals(self, theory=None, **kw):
-        for p, th in self.parts(theory):
-            p.plot_residuals(theory=th, **kw)
-
-    plot_residuals.__doc__ = Probe.plot_residuals.__doc__
 
     def parts(self, theory):
         if theory is None:
@@ -2059,108 +1751,6 @@ class PolarizedNeutronProbe:
                 xsi.save(filename + suffix, xsi_th, substrate=substrate, surface=surface)
 
     save.__doc__ = Probe.save.__doc__
-
-    def plot(self, view=None, **kwargs):
-        """
-        Plot theory against data.
-
-        Need substrate/surface for Fresnel-normalized reflectivity
-        """
-        view = view if view is not None else self.view
-
-        if view is None:
-            view = Probe.view  # Default to Probe.view
-
-        if view == "linear":
-            self.plot_linear(**kwargs)
-        elif view == "log":
-            self.plot_log(**kwargs)
-        elif view == "fresnel":
-            self.plot_fresnel(**kwargs)
-        elif view == "logfresnel":
-            self.plot_logfresnel(**kwargs)
-        elif view == "q4":
-            self.plot_Q4(**kwargs)
-        elif view.startswith("resid"):
-            self.plot_residuals(**kwargs)
-        elif view == "SA":
-            self.plot_SA(**kwargs)
-        elif view == "resolution":
-            self.plot_resolution(**kwargs)
-        else:
-            raise TypeError("incorrect reflectivity view '%s'" % view)
-
-    def plot_resolution(self, **kwargs):
-        self._xs_plot("plot_resolution", **kwargs)
-
-    def plot_linear(self, **kwargs):
-        self._xs_plot("plot_linear", **kwargs)
-
-    def plot_log(self, **kwargs):
-        self._xs_plot("plot_log", **kwargs)
-
-    def plot_fresnel(self, **kwargs):
-        self._xs_plot("plot_fresnel", **kwargs)
-
-    def plot_logfresnel(self, **kwargs):
-        self._xs_plot("plot_logfresnel", **kwargs)
-
-    def plot_Q4(self, **kwargs):
-        self._xs_plot("plot_Q4", **kwargs)
-
-    def plot_residuals(self, **kwargs):
-        self._xs_plot("plot_residuals", **kwargs)
-
-    def plot_SA(self, theory=None, label=None, plot_shift=None, **kwargs):
-        import matplotlib.pyplot as plt
-
-        if self.pp is None or self.mm is None:
-            raise TypeError("cannot form spin asymmetry plot without ++ and --")
-
-        plot_shift = plot_shift if plot_shift is not None else Probe.plot_shift
-        trans = auto_shift(plot_shift)
-        pp, mm = self.pp, self.mm
-        c = coordinated_colors()
-        if hasattr(pp, "R") and hasattr(mm, "R") and pp.R is not None and mm.R is not None:
-            Q, SA, dSA = spin_asymmetry(pp.Q, pp.R, pp.dR, mm.Q, mm.R, mm.dR)
-            if dSA is not None:
-                res = self.show_resolution if self.show_resolution is not None else Probe.show_resolution
-                dQ = pp.dQ if res else None
-                plt.errorbar(
-                    Q,
-                    SA,
-                    yerr=dSA,
-                    xerr=dQ,
-                    fmt=".",
-                    capsize=0,
-                    label=pp.label(prefix=label, gloss="data"),
-                    transform=trans,
-                    color=c["light"],
-                )
-            else:
-                plt.plot(Q, SA, ".", label=pp.label(prefix=label, gloss="data"), transform=trans, color=c["light"])
-            # Set limits based on max theoretical SA, which is in (-1.0, 1.0)
-            # If the error bars are bigger than that, you usually don't care.
-            ylim_low, ylim_high = plt.ylim()
-            plt.ylim(max(ylim_low, -2.5), min(ylim_high, 2.5))
-        if theory is not None:
-            mm, mp, pm, pp = theory
-            Q, SA, _ = spin_asymmetry(pp[0], pp[1], None, mm[0], mm[1], None)
-            plt.plot(Q, SA, label=self.pp.label(prefix=label, gloss="theory"), transform=trans, color=c["dark"])
-        plt.xlabel(r"Q (\AA^{-1})")
-        plt.ylabel(r"spin asymmetry $(R^{++} -\, R^{--}) / (R^{++} +\, R^{--})$")
-        plt.legend(numpoints=1)
-
-    def _xs_plot(self, plotter, theory=None, **kwargs):
-        import matplotlib.pyplot as plt
-
-        # Plot available cross sections
-        if theory is None:
-            theory = (None, None, None, None)
-        for x_data, x_th, suffix in zip(self.xs, theory, ("$^{--}$", "$^{-+}$", "$^{+-}$", "$^{++}$")):
-            if x_data is not None:
-                fn = getattr(x_data, plotter)
-                fn(theory=x_th, suffix=suffix, **kwargs)
 
 
 def spin_asymmetry(Qp, Rp, dRp, Qm, Rm, dRm):
